feat: initial container implementation

+++ src/service/config/source.rs
@@ -59,12 +59,7 @@ pub async fn validate(
         send_error(tx, "no source files found".to_string()).await?
     }
 
-    send(
-        tx,
-        format!("package source files: {}", workdir_files.len()),
-        None,
-    )
-    .await?;
+    send(tx, format!("source files: {}", workdir_files.len()), None).await?;
 
     let workdir_files_hashes = hashes::get_file_hashes(&workdir_files)?;
     let workdir_hash = hashes::get_source_hash(&workdir_files_hashes)?;
@@ -73,16 +68,11 @@ pub async fn validate(
         send_error(tx, "no source hash found".to_string()).await?
     }
 
-    send(
-        tx,
-        format!("package source hash: {}", workdir_hash),
-        None,
-    )
-    .await?;
+    send(tx, format!("source hash: {}", workdir_hash), None).await?;
 
     if let Some(request_hash) = &source.hash {
         if &workdir_hash != request_hash {
-            let message = &format!("hash mismatch: {} != {}", request_hash, workdir_hash);
+            let message = &format!("source hash mismatch: {} != {}", request_hash, workdir_hash);
             send_error(tx, message.to_string()).await?
         }
     }
@@ -93,20 +83,12 @@ pub async fn validate(
 pub async fn prepare(
     tx: &Sender<Result<ConfigPackageResponse, Status>>,
     source: &ConfigPackageSource,
-    source_tar_path: &PathBuf,
+    source_store_path: &PathBuf,
 ) -> Result<String, anyhow::Error> {
     let source_sandbox_path = temps::create_dir().await?;
-    let source_sandbox_path = source_sandbox_path.canonicalize()?;
-
-    let message = format!(
-        "package source sandbox: {:?}",
-        source_sandbox_path.file_name().unwrap()
-    );
-
-    send(tx, message.into(), None).await?;
 
     if source.kind == ConfigPackageSourceKind::Git as i32 {
-        let message = format!("package source git: {:?}", &source.uri);
+        let message = format!("source git: {:?}", &source.uri);
 
         send(tx, message.into(), None).await?;
 
@@ -160,82 +142,78 @@ pub async fn prepare(
         })
         .await?;
 
-        let message = format!("package source commit: {:?}", &result);
+        let message = format!("source commit: {:?}", &result);
 
         send(tx, message.into(), None).await?;
     }
 
     if source.kind == ConfigPackageSourceKind::Http as i32 {
-        let message = format!("package source url: {:?}", &source.uri);
+        let message = format!("source uri: {:?}", &source.uri);
 
         send(tx, message.into(), None).await?;
 
         let url = Url::parse(&source.uri)?;
 
         if url.scheme() != "http" && url.scheme() != "https" {
+            remove_dir_all(&source_sandbox_path).await?;
             send_error(tx, "invalid scheme".to_string()).await?;
         }
 
         let response = reqwest::get(url.as_str()).await?.bytes().await?;
+
         let response_bytes = response.as_ref();
 
-        let source_file_name = url.path_segments().unwrap().last().unwrap();
+        let file_name = url.path_segments().unwrap().last().unwrap();
 
         if let Some(kind) = infer::get(response_bytes) {
-            let message = format!("package source kind: {:?}", kind.mime_type());
+            let message = format!("source kind: {:?}", kind.mime_type());
 
             send(tx, message.into(), None).await?;
 
             if let "application/gzip" = kind.mime_type() {
                 let gz_decoder = GzipDecoder::new(response_bytes);
-
                 let mut archive = Archive::new(gz_decoder);
 
                 archive.unpack(&source_sandbox_path).await?;
 
-                let message = format!("package source gzip unpacked: {:?}", source_file_name);
+                let message = format!("source gzip unpacked: {:?}", file_name);
 
                 send(tx, message.into(), None).await?;
             } else if let "application/x-bzip2" = kind.mime_type() {
                 let bz_decoder = BzDecoder::new(response_bytes);
-
                 let mut archive = Archive::new(bz_decoder);
 
                 archive.unpack(&source_sandbox_path).await?;
 
-                let message = format!("package source bzip2 unpacked: {:?}", source_file_name);
+                let message = format!("source bzip2 unpacked: {:?}", file_name);
 
                 send(tx, message.into(), None).await?;
             } else if let "application/x-xz" = kind.mime_type() {
                 let xz_decoder = XzDecoder::new(response_bytes);
-
                 let mut archive = Archive::new(xz_decoder);
 
                 archive.unpack(&source_sandbox_path).await?;
 
-                let message = format!("package source xz unpacked: {:?}", source_file_name);
+                let message = format!("source xz unpacked: {:?}", file_name);
 
                 send(tx, message.into(), None).await?;
             } else if let "application/zip" = kind.mime_type() {
                 let temp_path = temps::create_file("zip").await?;
 
                 write(&temp_path, response_bytes).await?;
-
                 unpack_zip(&temp_path, &source_sandbox_path).await?;
-
                 remove_file(&temp_path).await?;
 
-                let message = format!("package source zip unpacked: {:?}", source_file_name);
+                let message = format!("source zip unpacked: {:?}", file_name);
 
                 send(tx, message.into(), None).await?;
             } else {
                 let file_name = url.path_segments().unwrap().last();
-
                 let sandbox_file_path = source_sandbox_path.join(file_name.unwrap());
 
                 write(&sandbox_file_path, response_bytes).await?;
 
-                let message = format!("package source file: {:?}", source_file_name);
+                let message = format!("source file: {:?}", file_name);
 
                 send(tx, message.into(), None).await?;
             }
@@ -244,18 +222,17 @@ pub async fn prepare(
 
     if source.kind == ConfigPackageSourceKind::Local as i32 {
         let source_path = Path::new(&source.uri).canonicalize()?;
-
-        let message = format!("package source local: {:?}", source_path);
+        let message = format!("source local: {:?}", source_path);
 
         send(tx, message.into(), None).await?;
 
         if let Ok(Some(source_kind)) = infer::get_from_path(&source_path) {
-            let message = format!("package source kind: {:?}", source_kind.mime_type());
+            let message = format!("source kind: {:?}", source_kind.mime_type());
 
             send(tx, message.into(), None).await?;
 
             if source_kind.mime_type() == "application/gzip" {
-                let message = format!("package source gzip unpacking: {:?}", source_path);
+                let message = format!("source gzip unpacking: {:?}", source_path);
 
                 send(tx, message.into(), None).await?;
 
@@ -269,7 +246,7 @@ pub async fn prepare(
             copy(&source_path, &dest).await?;
 
             let message = format!(
-                "package source file: {:?} -> {:?}",
+                "source file: {:?} -> {:?}",
                 source_path.display(),
                 dest.display()
             );
@@ -281,6 +258,7 @@ pub async fn prepare(
             let source_paths = get_file_paths(&source_path, &source.ignore_paths)?;
 
             if source_paths.is_empty() {
+                remove_dir_all(&source_sandbox_path).await?;
                 send_error(tx, "no source files found".to_string()).await?
             }
 
@@ -297,11 +275,7 @@ pub async fn prepare(
 
                 copy(path, &dest).await?;
 
-                let message = format!(
-                    "package source file: {:?} -> {:?}",
-                    path.display(),
-                    dest.display()
-                );
+                let message = format!("source file: {:?} -> {:?}", path.display(), dest.display());
 
                 send(tx, message.into(), None).await?;
             }
@@ -311,28 +285,25 @@ pub async fn prepare(
     // At this point, any source URI should be a local file path
 
     let (source_sandbox_hash, _) = validate(tx, &source_sandbox_path, source).await?;
-
     let source_sandbox_file_paths = get_file_paths(&source_sandbox_path, &Vec::<&str>::new())?;
 
-    let message = format!("sandbox source files: {}", source_sandbox_file_paths.len());
-
-    send(tx, message.into(), None).await?;
+    if source_sandbox_file_paths.is_empty() {
+        remove_dir_all(&source_sandbox_path).await?;
+        send_error(tx, "no source files found".to_string()).await?
+    }
 
     compress_zstd(
         &source_sandbox_path,
         &source_sandbox_file_paths,
-        source_tar_path,
+        source_store_path,
     )
     .await?;
 
-    let message = format!(
-        "package source archive: {:?}",
-        source_tar_path.file_name().unwrap()
-    );
+    remove_dir_all(&source_sandbox_path).await?;
+
+    let message = format!("source store: {:?}", source_store_path.file_name().unwrap());
 
     send(tx, message.into(), None).await?;
-
-    remove_dir_all(&source_sandbox_path).await?;
 
     Ok(source_sandbox_hash)
 }
